moved trailing buttons widget to new file
