+++ lib/notes/list_notes.dart
@@ -31,13 +31,12 @@ import 'package:notepod/constants/ui.dart';
 import 'package:notepod/models/note.dart';
 import 'package:notepod/models/selected_note.dart';
 import 'package:notepod/notes/non_readable_note.dart';
-import 'package:notepod/notes/share_note.dart';
 import 'package:notepod/notes/view_note.dart';
 import 'package:notepod/shared_notes/list_external_notes_screen.dart';
 import 'package:notepod/utils/get_id.dart';
 import 'package:notepod/utils/misc.dart';
+import 'package:notepod/widgets/note_item_trailing_buttons.dart';
 import 'package:notepod/widgets/note_list_del_button.dart';
-import 'package:notepod/widgets/simple_action_button.dart';
 
 /// A [stateful] widget to list notes accessible to the
 /// user.
@@ -611,7 +610,7 @@ class _ListNotesState extends State<ListNotes> {
                             trailing: SizedBox(
                               height: 60,
                               width: 120,
-                              child: SharedTrailingButtons(
+                              child: NoteItemTrailingButtons(
                                 note: _foundNotes[index],
                                 scaffoldController: _scaffoldController,
                               ),
@@ -647,48 +646,6 @@ class _ListNotesState extends State<ListNotes> {
           ),
         );
       },
-    );
-  }
-}
-
-class SharedTrailingButtons extends StatelessWidget {
-  const SharedTrailingButtons({
-    super.key,
-    required Note note,
-    required SolidScaffoldController scaffoldController,
-  })  : _note = note,
-        _scaffoldController = scaffoldController;
-
-  final Note _note;
-  final SolidScaffoldController _scaffoldController;
-
-  @override
-  Widget build(BuildContext context) {
-    List accessList = _note.permissionList.split(',');
-
-    return Row(
-      mainAxisAlignment: MainAxisAlignment.end,
-      spacing: 5.0,
-      children: [
-        // Share button if control in permissions
-        if (accessList.contains('control')) ...[
-          SimpleActionButton(
-            icon: const Icon(Icons.share),
-            childPage: ShareNote(
-              noteUrl: _note.noteUrl,
-              noteOwner: _note.noteOwner,
-              isExternal: _note.isExternalRes,
-              backPage: ListExternalNotesScreen(
-                scaffoldController: _scaffoldController,
-              ),
-              scaffoldController: _scaffoldController,
-            ),
-            scaffoldController: _scaffoldController,
-          ),
-        ],
-        // Open note icon
-        const Icon(Icons.arrow_forward),
-      ],
     );
   }
 }
